Refactor eta and zeta tower helpers
